// src/app/interfaces/Serie.ts
import { Personal } from "./Anime";

export interface SerieResult {
    page:          number;
    results:       Serie[];
    total_pages:   number;
    total_results: number;
}

export interface Serie {
    adult:             boolean;
    backdrop_path:     string;
    genre_ids:         number[];
    id:                number;
    origin_country:    OriginCountry[];
    original_language: OriginalLanguage;
    original_name:     string;
    overview:          string;
    popularity:        number;
    poster_path:       string;// Generated by https://quicktype.io


    first_air_date:    Date;
    name:              string;
    vote_average:      number;
    vote_count:        number;
    isSaved?: boolean;
    personal?: Personal;
}

export enum OriginCountry {
    CA = "CA",
    Jp = "JP",
    Kr = "KR",
    Us = "US",
}

export enum OriginalLanguage {
    En = "en",
    Ja = "ja",
    Ko = "ko",
}

export interface GenreSeries {
    genres: Genre[];
}

export interface Genre {
    id:   number;
    name: string;
}